Add review JWT tokken

+++ part3/app/api/v1/reviews.py
@@ -17,17 +17,31 @@ class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @api.response(400, 'You cannot review your own place.')
+    @api.response(400, 'You have already reviewed this place.')
+    @jwt_required()
     def post(self):
         """Register a new review"""
         review_data = api.payload
+        current_user = get_jwt_identity()
+
+        user = facade.get_by_attribute('id', current_user)
+        if not user:
+            return {'error': 'User not found'}, 400
+
         place = facade.get_place(review_data['place_id'])
         if not place:
             return {'error': 'Place not found'}, 400
-        user = facade.get_user(review_data['user_id'])
-        if not user:
-            return {'error': 'User not found'}, 400
-        if place.owner.id == user.id:
-            return {'error': 'User cannot review their own place'}, 400
+        
+        if place.owner_id == current_user:
+            return {'error': 'You cannot review your own place.'}, 400
+
+        existing_reviews = facade.get_reviews_by_place(review_data['place_id'])
+        for review in existing_reviews:
+            if review.user.id == current_user:
+                return {'error': 'You have already reviewed this place.'}, 400
+        review_data['user_id'] = current_user
+
         try:
             new_review = facade.create_review(review_data)
             return new_review.to_dict(), 201
@@ -54,12 +68,18 @@ class ReviewResource(Resource):
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def put(self, review_id):
         """Update a review's information"""
         review_data = api.payload
+        current_user = get_jwt_identity()
+        
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
+        
+        if review.user_id != current_user:
+            return {'error': 'Unauthorized action'}, 403
         
         try:
             facade.update_review(review_id, review_data)
@@ -69,12 +89,18 @@ class ReviewResource(Resource):
 
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @jwt_required()
     def delete(self, review_id):
         """Delete a review"""
+        current_user = get_jwt_identity()
+
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
-        
+
+        if review.user_id != current_user:
+            return {'error': 'Unauthorized action'}, 403
+
         try:
             facade.delete_review(review_id)
             return {'message': 'Review deleted successfully'}, 200
